check 2

+++ so_long/src/coin.c
@@ -6,7 +6,7 @@
 /*   By: haryu <haryu@student.42seoul.co.kr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/23 12:46:10 by haryu             #+#    #+#             */
-/*   Updated: 2022/03/24 20:55:30 by haryu            ###   ########.fr       */
+/*   Updated: 2022/03/25 16:06:07 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ int	check_x_y_coin(t_module *init, char **map)
 	int	j;
 
 	i = 0;
-	while (i < init->map.position.y / SIZE)
+	while (i < (int)init->map.position.y / SIZE)
 	{
 		j = 0;
 		while (map[i][j] != '\n' && map[i][j] != '\0')
